+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Dynamic Iframe component using route parameters to drive rendered iframe src
+
 ## [0.0.4] - 2019-10-18
 
 ## [0.0.3] - 2019-10-17